Reverse attribute access for creating the data object.
Mongoose is used as an example. Creating a document object in mongoose
gives the 'record' object more than just the attributes to serialize. In
fact, the attributes we're looking for are in the `_doc` property of the
document object. I'm sure this pattern is not just specific to mongoose,
although I don't have enough experience with another O(D|R)M module to
say for certain.

The previous attribute access scheme iterated through the object's keys
to see if the key was in the option's attributes array. This
meant that it did not recurse to find the attributes in `_doc` and
didn't allow the passed record to figure out how to handle the calls.
Additionally, for very large objects, I can imagine this becoming slow.
We're taking two sets and looking for intersection using the larger
set first and checking for existance in the smaller set.

But don't believe me, I haven't done any optimizations like this before.

The new attribute access scheme iterates through the smaller set, the
option's attributes array, and asking the record object for that
attribute. This allows the record object to figure out how to handle it
using all of the magic that it's creating module gave it.

There is also a new test that confirms this behavior.

Fixes #52.