Addd support for a static overriding of relationship types.
This means that if you want to have a simple typename for a relationship record, you can do so by simply specifying it in the config without needing to use the typeForAttribute confguration. typeForAttribute will still overwrite this value so that you can dynamically change the attribute type in some situations if needed.